requirements.txt
Pillow
numpy